Commented nand2tetris.java and gui.java
